Add flv_reader.

+++ protocol/libflv/include/flv.h
@@ -10,7 +10,7 @@
 
 typedef struct flv_handler_t
 {
-    int (*on_write)(void *param, const void *data, size_t bytes);
+    int (*on_write)(void *param, const void *data, uint32_t bytes);
 
     /// @param[in] param user-defined parameter
     /// @param[in] type audio/video format (see flv_type.h)
@@ -18,7 +18,7 @@ typedef struct flv_handler_t
     /// @param[in] bytes data length
     /// @param[in] timestamp timestamp in milliseconds
     /// @return 0-ok, other-error
-    int (*on_frame_write)(void* param, int type, const void* data, size_t bytes, uint32_t timestamp);
+    int (*on_frame_write)(void* param, int type, const void* data, uint32_t bytes, uint32_t timestamp);
 
     /// @param[in] param user-defined parameter
     /// @param[in] codec audio/video format (see flv_type.h)
@@ -28,13 +28,13 @@ typedef struct flv_handler_t
     /// @param[in] dts audio/video decoding timestamp
     /// @param[in] key 1-video keyframe, other-undfined
     /// @return 0-ok, other-error
-    int (*on_packet)(void* param, int codec, const void* data, size_t bytes, uint32_t pts, uint32_t dts, int key);
+    int (*on_packet)(void* param, int codec, const void* data, uint32_t bytes, uint32_t pts, uint32_t dts, int key);
 } flv_handler_t;
 
 typedef struct flv_vec_t
 {
     const void *data;
-    size_t      bytes;
+    uint32_t    bytes;
 } flv_vec_t;
 
 #endif //LIBFLV_FLV_H
